PasswordManager/requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
Naked==0.1.32
pillow==10.3.0
pycryptodome==3.20.0
PyYAML==6.0.1
requests==2.31.0
shellescape==3.8.1
urllib3==2.2.1
